+++ src/test/java/step3/domain/round/NumberOfRoundsTest.java
@@ -1,4 +1,4 @@
-package step3.domain.game;
+package step3.domain.round;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
